A fix that addresses an apparant cause of the issues that were being had with updates resulting backpropogation steps resulting in values that were updated in the example network becoming nan. In particular, using the same fix as noted in commit 413c7d07d4705fef121cadf2c16b5de7823cf602 (different branch) of no longer using "half-type" floats worked. For record, the initial impetus to do that was that the earlier versions of the example network that were larger were having trouble fitting on the GPU with the full-sized datatypes.

+++ src/models/components/simple_dense_net.py
@@ -36,8 +36,8 @@ class SimpleDenseNet(nn.Module):
         )
         """
         self.model = nn.Sequential(
-            nn.Linear(input_size, lin1_size,dtype=torch.float16),
-            nn.Linear(lin1_size, output_size,dtype=torch.float16)
+            nn.Linear(input_size, lin1_size),
+            nn.Linear(lin1_size, output_size)
         );
         for index in [0,1]:
             self.model[index].bias.data = 0 * self.model[index].bias.data;
